+++ js/index.js
@@ -1,4 +1,4 @@
-import * as templatesLib from "./templates/templates.js";
+import * as templatesLib from "./templates/template_lib.js";
 import {config} from "./config.js";
 import * as editor from "./editor/editor.js";
 import * as console from "./console/console.js";
@@ -7,6 +7,7 @@ import * as console from "./console/console.js";
 
 // Templates
 templatesList.addEventListener("click", templateSelected);
+templatesCategories.addEventListener("click", templateCategorySelected);
 
 // Search in Templates
 document.querySelector("#wed-search").addEventListener("change", templatesLib.filterTemplates); // Paste
@@ -16,8 +17,9 @@ document.querySelector("#wed-search").addEventListener("keyup", templatesLib.fil
 window.addEventListener('load', init);
 
 function init() {  
-  console.consoleHook("textareaConsole");
-  templatesLib.initializeTemplates();
+  console.consoleHook('textareaConsole');
+  templatesLib.setupCategoriesContainer('templatesCategories');
+  templatesLib.setupContainer('templatesList');
   config.version_check();
   if (editor.init(config)) {
     getTemplate( templatesLib.templates[0].files);
@@ -27,26 +29,54 @@ function init() {
 
 // Templates
 
+function templateCategorySelected(event) {
+  if (! event.target.matches("button")) 
+        return;
+
+  templatesLib.setupContainer('templatesList', event.target.innerText);
+}
+
+/*
+templates: [
+
+    category
+    ↓
+  'html': [
+
+    category items
+        ↓
+    { "name": "cards", "lang": "html", "files": "cards.html , cards.css" },
+    { "name": "progress", "lang": "html", "files": "progress.html , progress.css" }
+     …
+   ],
+   'css': …
+
+]
+*/
 function templateSelected(event) {
     
+    // event delegation. Target div childs
     if (! event.target.matches("div")) 
         return;
     
     const content = event.srcElement.getAttribute("data-content");          
     if (content != null) {
-        applyTemplate(null,content);
+        renderTemplate(null,content);
+        return;
     }
-    else {
-        const files = event.srcElement.getAttribute("data-files");   
-      getTemplate( files );
-    }
+
+    const files = event.srcElement.getAttribute("data-files");   
+    const category = event.srcElement.getAttribute("data-category");   
+    getTemplate( files, category );
 }    
 
-function getTemplate(files) {
-  templatesLib.getTemplateFiles(config.templatesDirectory, files, applyTemplate);          
+function getTemplate(files, category) {
+  templatesLib.getTemplate(`${config.templatesDirectory}/${category}`, files, renderTemplate);          
 }
 
-function applyTemplate(file, result) {
+function renderTemplate(file, result) {
+  
+  // template file not found
   if (file == null) {
     editor.cmCss.setValue("");  
     editor.cmHtml.setValue(result);     
@@ -54,9 +84,10 @@ function applyTemplate(file, result) {
   } 
   
   const ext = file.split(".");
-
   if (ext.length < 2) {
-    console.log(`Error with ${file}: no extension`);
+    const msg = `Error with ${file}: no extension`;
+    editor.cmHtml.setValue(msg);
+    console.log(msg);
     return;
   }