+++ js/utilities/utilities.js
@@ -6,3 +6,12 @@ export function httpGet(url, callback=console.info, err = console.error) {
   request.onerror = () => err(request);
   request.send();
 }
+
+export async function copyToClipboard(stringToCopy, show = null) {
+  try {
+    await navigator.clipboard.writeText(stringToCopy.toString());
+    // this.snackbar(`copied ${show == true ? stringToCopy : ""}`);
+  } catch (err) {
+    console.error(`Failed to copy ${stringToCopy}`, err);
+  }
+}